test: move test files into __tests__ directories
